requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-cors-headers==2.2.0
djangorestframework==3.7.7
idna==2.6
pytz==2018.3
requests==2.18.4
unicodecsv==0.14.1
urllib3==1.22
